deprecated original `index.ts`
