+++ deprecated/README.md
@@ -1,0 +1,4 @@
+# Deprecated Stuff
+This folder mostly contains deprecated and old files that will be removed in the future
+
+please use the newer files instead.
